+++ tests/opc/test_package.py
@@ -154,7 +154,9 @@ class DescribeOpcPackage:
         pkg = OpcPackage()
         pkg.save(pkg_file_)
         for part in parts_:
-            part.before_marshal.assert_called_once_with(reproducible=False)
+            part.before_marshal.assert_called_once_with(
+                reproducible=False, mirror_paragraph_marks=False,
+            )
         PackageWriter_.write.assert_called_once_with(
             pkg_file_, pkg.rels, parts_, reproducible=False, password=None
         )